Task-09/pokedex/api/requirements.txt
python-dotenv==1.1.0
aiodns==3.1.1
aiohttp==3.9.1
aiosignal==1.3.1
appdirs==1.4.4
argcomplete==3.1.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==22.7.1
Automat==22.10.0
Babel==2.10.3
base58==1.0.3
bcc==0.29.1
bcrypt==3.2.2
beautifulsoup4==4.12.3
blinker==1.7.0
Bottleneck==1.3.5
Brlapi==0.8.5
Brotli==1.1.0
cbor==1.0.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
chardet==5.2.0
click==8.1.6
cloud-init==24.4.1
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
contourpy==1.0.7
cryptography==41.0.7
cssselect==1.2.0
cupshelpers==1.0
cycler==0.11.0
daphne==4.0.0
dbus-python==1.3.2
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.9.0
distro-info==1.7+build1
Django==4.2.11
ecdsa==0.18.0
et-xmlfile==1.0.1
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.3
flatbuffers==2.0.8+dfsg1
fonttools==4.46.0
frozenlist==1.4.0
fs==2.4.16
greenlet==3.0.3
gyp==0.1
html5lib==1.1
httplib2==0.20.4
hyperlink==21.0.0
idna==3.6
importlib-metadata==4.12.0
incremental==22.10.0
itsdangerous==2.1.2
jaraco.classes==3.2.1
jeepney==0.8.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
keyring==24.3.1
kiwisolver==0.0.0
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
lxml==5.2.1
lz4==4.0.2+dfsg
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.6.3
mdurl==0.1.2
mnemonic==0.19
more-itertools==10.2.0
mpmath==0.0.0
msgpack==1.0.3
multidict==6.0.4
netaddr==0.8.0
numexpr==2.9.0
numpy==1.26.4
oauthlib==3.2.2
odfpy==1.4.2
olefile==0.46
openpyxl==3.1.2
packaging==24.0
pandas==2.1.4+dfsg
passlib==1.7.4
pexpect==4.9.0
pillow==10.2.0
pipx==1.4.3
platformdirs==4.2.0
ply==3.11
proton-core==0.4.0
proton-keyring-linux==0.2.0
proton-vpn-api-core==0.42.4
proton-vpn-gtk-app==4.9.5
proton-vpn-network-manager==0.12.13
psutil==5.9.8
ptyprocess==0.7.0
py-cpuinfo==9.0.0
py-ubjson==0.16.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyasyncore==1.0.2
pycairo==1.25.1
pycares==4.4.0
pycparser==2.21
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
pyinotify==0.9.6
PyJWT==2.7.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pypng==0.20231004.0
PyQRCode==1.2.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-gnupg==0.5.2
python-snappy==0.5.3
PyTrie==0.4.0
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
SciPy==1.11.4
screen-resolution-extra==0.0.0
SecretStorage==3.3.3
sentry-sdk==1.39.2
service-identity==24.1.0
setuptools==68.1.2
simplejson==3.19.2
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.50
sqlparse==0.4.4
sympy==1.12
systemd-python==235
tables==3.9.2
tornado==6.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
u-msgpack-python==2.3.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.16.0
ufw==0.36.2
ujson==5.9.0
unattended-upgrades==0.1
unicodedata2==15.1.0
urllib3==2.0.7
userpath==1.9.1
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==3.0.1
wheel==0.42.0
wsaccel==0.6.3
xdg==5
xkit==0.0.0
yarl==1.8.2
zipp==1.0.0
zope.interface==6.1
